+++ lib/features/branches/presentation/bloc/branches_state.dart
@@ -7,6 +7,7 @@ enum BranchesBlocStatus {
   error,
   createNewBranchAndCheckout,
   branchCreated,
+  branchesRetrieved,
 }
 
 @MappableClass()